Add source file options to maintain transparency (#7)
Fixes converted pngs having a white background added to them when uploading an svg.

Turns out this argument has to be set before the file input in the convert command.
This is more or less a hotfix, which works for now.
Doesn't mess with other upload types.

A better solution would be to have different options for different file types, as `media_attachment.rb` has to also fix images not scaling to the required size.

Signed-off-by: PolyamSpace <117664621+polyamspace@users.noreply.github.com>