Skip foreign key migration when correct keys already exist
I'm not sure how this migration behaves when combined with the upstream PR. Best case it just unecessarily recreates the foreign keys, but I have a feeling it would error out instead.

This change essentially makes this migration do nothing when upstream's migration was run first.

Also renames the migration, so it's clearer what it does.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>